Update and rename REG-SEC-GPO RAW Script to REG-SEC-GPO_RAW-Script
Version 1.1 Update:

VISUAL CHANGES:
- Sub-menus have their own title void of other title names.
IE. In "REG", present is "REG", not "REG-SEC-GPO".

- All "v 1.0" updated to display "v1.1" update.

CODE CHANGES:
- Following Sub-Menus created:
-- REG
-- SEC (Placeholder for now)
-- GPO

All previous functionalities haven't changed, just now updated to be in a nicer format. 

Each sub-menu now displays their own area specific choices, allowing for more clarity and separation. 

Whereby, REG tools are in the sub-menu "REG" and SEC tool (will) be in the sub-menu "SEC" and lastly, GPO tools will be in the "GPO" sub-menu.

"Installer" is still present on "Top Menu", as this is not specific to a separate category.